+++ pom.xml
@@ -33,6 +33,7 @@
         <module>plugins</module>
         <module>dashboard</module>
         <module>site</module>
+        <module>ide/codenvy-hosted-basic-ide</module>
         <module>assembly</module>
         <module>dockerfiles/agents</module>
     </modules>
@@ -207,6 +208,11 @@
             </dependency>
             <dependency>
                 <groupId>com.codenvy.onpremises</groupId>
+                <artifactId>codenvy-hosted-basic-ide</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.codenvy.onpremises</groupId>
                 <artifactId>compiling-ide-war</artifactId>
                 <version>${project.version}</version>
                 <classifier>classes</classifier>
@@ -223,6 +229,12 @@
                 <version>${project.version}</version>
                 <type>zip</type>
                 <classifier>${site.classifier}</classifier>
+            </dependency>
+            <dependency>
+                <groupId>com.codenvy.onpremises.wsagent</groupId>
+                <artifactId>codenvy-hosted-basic-server</artifactId>
+                <version>${project.version}</version>
+                <type>war</type>
             </dependency>
             <dependency>
                 <groupId>com.codenvy.onpremises.wsagent</groupId>
