+++ wsagent/codenvy-hosted-basic-server/src/main/java/com/codenvy/wsagent/server/WsAgentAnalyticsAddresser.java
@@ -12,7 +12,7 @@
  * is strictly forbidden unless prior written permission is obtained
  * from Codenvy S.A..
  */
-package com.codenvy.ext.java.server;
+package com.codenvy.wsagent.server;
 
 import org.eclipse.che.commons.lang.NameGenerator;
 import org.eclipse.che.commons.schedule.ScheduleRate;
