+++ wsagent/codenvy-hosted-basic-server/src/main/webapp/WEB-INF/classes/codenvy/che-machine-configuration.properties
