+++ wsagent/codenvy-hosted-basic-server/src/main/webapp/WEB-INF/web.xml
