+++ wsagent/pom.xml
@@ -31,5 +31,6 @@
         <module>wsagent-codenvy</module>
         <module>codenvy-hosted-machine-authentication-agent</module>
         <module>codenvy-hosted-workspace-activity-agent</module>
+        <module>codenvy-hosted-basic-server</module>
     </modules>
 </project>
